refactor(web/stores): split settings into package
Decomposes the 509-line stores/settings.ts monolith into a
sibling-aggregator package. The mutation-token + entries-generation
concurrency logic is preserved verbatim; the 15 characterisation
tests landed in commit 1 stay green throughout.

- stores/settings.ts (38 lines): aggregator. Hosts the
  updateFromWsEvent action inline (it's a one-line system-channel
  refreshEntries trigger).
- stores/settings/types.ts: SettingsState plus SettingsSet/SettingsGet.
- stores/settings/concurrency.ts: incrementSavingKey,
  decrementSavingKey, nextMutationToken (the module-scoped counter),
  deriveCurrency, fetchAllSettingsEntries (paginated walker),
  CURRENCY_PATTERN. These are the shared primitives the fetch and
  mutation slices both rely on.
- stores/settings/fetch-actions.ts: fetchCurrency, fetchSettingsData,
  refreshEntries. Each is a top-level *Impl function so the slice
  factory body stays under 80 LPF.
- stores/settings/mutation-actions.ts: updateSetting, resetSetting.
  The success / out-of-order / error paths and the post-reset
  generation-drift apply logic are decomposed into named helpers
  (applyUpdatedEntry, emitOutOfOrderDrop, handleUpdateError,
  fetchAfterReset, applyResetSnapshot, buildResetPatch,
  applyFreshSnapshot) so every function stays under cx 8 and LPF 80.

Every new file under 400 LOC, every function under 80 LPF, every
function complexity at most 8. All 15 settings characterisation
tests pass unchanged.

Part of #2093 (PR B of EPIC #2066).
